+++ crates/test-programs/wasi-tests/src/bin/stdio_isatty.rs
@@ -1,22 +1,22 @@
-unsafe fn test_stdio_not_isatty() {
+unsafe fn test_stdio_isatty() {
     assert_eq!(
         libc::isatty(libc::STDIN_FILENO as std::os::raw::c_int),
-        0,
-        "stdin is not a tty"
+        1,
+        "stdin is a tty"
     );
     assert_eq!(
         libc::isatty(libc::STDOUT_FILENO as std::os::raw::c_int),
-        0,
-        "stdout is not a tty"
+        1,
+        "stdout is a tty"
     );
     assert_eq!(
         libc::isatty(libc::STDERR_FILENO as std::os::raw::c_int),
-        0,
-        "stderr is not a tty"
+        1,
+        "stderr is a tty"
     );
 }
 
 fn main() {
     // Run the tests.
-    unsafe { test_stdio_not_isatty() }
+    unsafe { test_stdio_isatty() }
 }
