+++ crates/wasi-common/cap-std-sync/Cargo.toml
@@ -23,7 +23,7 @@ fs-set-times = { workspace = true }
 system-interface = { workspace = true, features = ["cap_std_impls"] }
 tracing = { workspace = true }
 io-lifetimes = { workspace = true }
-is-terminal = "0.4.0"
+is-terminal = { workspace = true }
 
 [target.'cfg(unix)'.dependencies]
 rustix = { workspace = true, features = ["fs", "event"] }
